[IMP] pos*: move customer_display_adapter to pos/app
In this commit:
================
Moved the customer_display_adapter file from the customer_display folder to
pos/app/customer_display because it contains logic that is directly used by the
POS app. The principle is to keep all code used by the POS app within the
pos/app directory.

This change helps maintain a consistent structure where all dependencies used
in POS are centralized in its app directory.
