+++ addons/l10n_in_pos/__manifest__.py
@@ -23,7 +23,6 @@
         'point_of_sale._assets_pos': [
             'l10n_in/static/src/helpers/hsn_summary.js',
             'l10n_in_pos/static/src/**/*',
-            'l10n_in_pos/static/src/customer_display/customer_display_adapter.js',
         ],
     },
     'author': 'Odoo S.A.',
