+++ addons/l10n_in_pos/static/src/app/customer_display/customer_display_adapter.js
@@ -1,5 +1,5 @@
 import { patch } from "@web/core/utils/patch";
-import { CustomerDisplayPosAdapter } from "@point_of_sale/customer_display/customer_display_adapter";
+import { CustomerDisplayPosAdapter } from "@point_of_sale/app/customer_display/customer_display_adapter";
 
 patch(CustomerDisplayPosAdapter.prototype, {
     formatOrderData(order) {
