+++ addons/point_of_sale/static/src/app/customer_display/customer_display_adapter.js
