+++ addons/point_of_sale/static/src/app/pos_app.js
@@ -6,7 +6,7 @@ import { reactive, Component, onMounted, onWillStart } from "@odoo/owl";
 import { effect } from "@web/core/utils/reactive";
 import { batched } from "@web/core/utils/timing";
 import { useOwnDebugContext } from "@web/core/debug/debug_context";
-import { CustomerDisplayPosAdapter } from "@point_of_sale/customer_display/customer_display_adapter";
+import { CustomerDisplayPosAdapter } from "@point_of_sale/app/customer_display/customer_display_adapter";
 import { useIdleTimer } from "./utils/use_idle_timer";
 import useTours from "./hooks/use_tours";
 import { init as initDebugFormatters } from "./utils/debug-formatter";
